+++ .gitignore
@@ -5,6 +5,12 @@
 /tmp
 /out-tsc
 
+# env
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
+
 # dependencies
 /node_modules
