--- apps/api/src/app/app.service.ts
@@ -1,8 +1,0 @@
-import { Injectable } from '@nestjs/common'
-
-@Injectable()
-export class AppService {
-  getData(): { message: string } {
-    return { message: 'Welcome to my-nest-app!' }
-  }
-}
