--- apps/api/src/main.ts
@@ -1,26 +1,0 @@
-/**
- * This is not a production server yet!
- * This is only a minimal backend to get started.
- */
-
-import { Logger } from '@nestjs/common'
-import { NestFactory } from '@nestjs/core'
-
-import { AppModule } from './app/app.module'
-
-import { google } from 'googleapis'
-
-async function bootstrap() {
-  const test = google._options
-  console.log(test.baseURL)
-  const app = await NestFactory.create(AppModule)
-  const globalPrefix = 'api'
-  app.setGlobalPrefix(globalPrefix)
-  const port = process.env.PORT || 3333
-  await app.listen(port)
-  Logger.log(
-    `🚀 Application is running on: http://localhost:${port}/${globalPrefix}`
-  )
-}
-
-bootstrap()
