+++ apps/gdrive-s3-sync-api/Dockerfile
@@ -1,9 +1,8 @@
 FROM node:lts-alpine
 WORKDIR /apps
-COPY ./dist/apps/api .
+COPY ./dist/apps/gdrive-s3-sync-api .
 ENV PORT=3333
 EXPOSE ${PORT}
 RUN npm install --production
 # dependencies that nestjs needs
-RUN npm install reflect-metadata tslib rxjs @nestjs/platform-express
 CMD node ./main.js
