+++ apps/gdrive-s3-sync-api/jest.config.js
@@ -1,5 +1,5 @@
 module.exports = {
-  displayName: 'api',
+  displayName: 'gdrive-s3-sync-api',
   preset: '../../jest.preset.js',
   globals: {
     'ts-jest': {
@@ -11,5 +11,5 @@ module.exports = {
     '^.+\\.[tj]s$': 'ts-jest',
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/apps/api',
+  coverageDirectory: '../../coverage/apps/gdrive-s3-sync-api',
 };
