+++ apps/gdrive-s3-sync-api/src/app/app.controller.spec.ts
@@ -1,24 +1,24 @@
-import { Test, TestingModule } from '@nestjs/testing'
+import { Test, TestingModule } from '@nestjs/testing';
 
-import { AppController } from './app.controller'
-import { AppService } from './app.service'
+import { AppController } from './app.controller';
+import { AppService } from './app.service';
 
 describe('AppController', () => {
-  let app: TestingModule
+  let app: TestingModule;
 
   beforeAll(async () => {
     app = await Test.createTestingModule({
       controllers: [AppController],
       providers: [AppService],
-    }).compile()
-  })
+    }).compile();
+  });
 
   describe('getData', () => {
     it('should return "Welcome to my-nest-app!"', () => {
-      const appController = app.get<AppController>(AppController)
+      const appController = app.get<AppController>(AppController);
       expect(appController.getData()).toEqual({
         message: 'Welcome to my-nest-app!',
-      })
-    })
-  })
-})
+      });
+    });
+  });
+});
