+++ apps/gdrive-s3-sync-api/src/app/app.service.ts
@@ -1,0 +1,8 @@
+import { Injectable } from '@nestjs/common';
+
+@Injectable()
+export class AppService {
+  getData(): { message: string } {
+    return { message: 'Welcome to my-nest-app!' };
+  }
+}
