+++ apps/gdrive-s3-sync-api/src/environments/environment.prod.ts
@@ -1,3 +1,3 @@
 export const environment = {
-  production: false,
-}
+  production: true,
+};
