+++ apps/gdrive-s3-sync-api/src/environments/environment.ts
@@ -1,3 +1,3 @@
 export const environment = {
-  production: true,
-}
+  production: false,
+};
