+++ apps/gdrive-s3-sync-api/tsconfig.spec.json
